tools/requirements.txt
grpcio-tools==1.51.1
pydantic==2.8.2
fastapi==0.112.0
uvicorn==0.22.0
sqlalchemy==1.4.48
sqlalchemy-stubs==0.4
pythonnet==3.0.1
requests==2.29.0 
types-requests==2.28.11.17
ruff==0.0.256
mypy==1.11.2
mypy-protobuf==3.4.0
types-protobuf==4.22.0.0
types_pywin32==305.0.0.11
types_pyserial==3.5.0.7
grpc-stubs==1.24.12.1
networkx==3.1
networkx-stubs==0.0.1
types-setuptools==68.0.0.2
types-PyYAML==6.0.12.11
PyYAML==6.0.1
slackclient==2.9.4
python-dotenv==1.0.0
pytest==7.4.3
slack-sdk==3.21.3
msl-loadlib==0.10.0
pywin32==304; platform_system == "Windows"
#markdown==3.6
#pdfkit==1.0.0
pillow==10.3.0
typing-inspect==0.8.0
typing_extensions==4.8.0